+++ cloud-annotations/src/main/java/cloud/commandframework/annotations/CommandMethod.java
@@ -35,6 +35,7 @@ import org.checkerframework.checker.nullness.qual.NonNull;
 @Retention(RetentionPolicy.RUNTIME)
 @Target({ElementType.METHOD, ElementType.TYPE})
 public @interface CommandMethod {
+    String ANNOTATION_PATH = "cloud.commandframework.annotations.CommandMethod";
 
     /**
      * Command syntax