+++ cloud-annotations/src/test/resources/TestCommandMethodPrivate.java
@@ -1,0 +1,12 @@
+import cloud.commandframework.annotations.Argument;
+import cloud.commandframework.annotations.CommandMethod;
+
+public class TestCommandMethodPrivate {
+
+    @CommandMethod("command <required> [optional]")
+    private void commandMethod(
+            @Argument("required") final String required,
+            @Argument("optional") final String optional
+    ) {
+    }
+}